Split SVG.php into Length, Angle, Color, ColorLookup
